create database directory, move seeds dir into database dir
